fix: re-introduce size_t and re-write syscall as utility funcs

+++ engine/src/common/constants.h
@@ -9,18 +9,18 @@
 #include <cstdint>
 
 // Embedding vector dimension produced by `all-MiniLM-L6-v2`.
-inline constexpr uint32_t kVectorDim = 384;
+inline constexpr size_t kVectorDim = 384;
 // The corresponding byte footprint of one vector.
-inline constexpr uint32_t kVectorMemsize = kVectorDim * sizeof(float);
+inline constexpr size_t kVectorMemsize = kVectorDim * sizeof(float);
 
 // Minimum cosine similarity for a cache hit.
 inline constexpr float kSimilarityThreshold = 0.85f;
 
 // L0 buffer's slot limits.
-inline constexpr uint32_t kL0Capacity = 1 << 12;
+inline constexpr size_t kL0Capacity = 1 << 12;
 // Total number of slots in the unified `MemoryArena`.
-inline constexpr uint32_t kTotalSlots = 1 << 19;
+inline constexpr size_t kTotalSlots = 1 << 19;
 
 // Maximum lifetime of a `kPending` node in seconds.
 // Nodes remain `kPending` beyond this deadline are treated as stale by the GC.
-inline constexpr uint32_t kPendingLifespan = 30;
+inline constexpr uint64_t kPendingLifespan = 30;
